+++ packages/abap-deploy-config-sub-generator/src/app/index.ts
@@ -14,11 +14,17 @@ import {
 } from '@sap-ux/fiori-generator-shared';
 import { getPackageAnswer, getTransportAnswer, reconcileAnswers } from '@sap-ux/abap-deploy-config-inquirer';
 import { generate as generateAbapDeployConfig } from '@sap-ux/abap-deploy-config-writer';
-import { initTelemetrySettings } from '@sap-ux/telemetry';
 import { UI5Config } from '@sap-ux/ui5-config';
 import { FileName, getAppType } from '@sap-ux/project-access';
 import { AuthenticationType } from '@sap-ux/store';
-import { t, handleProjectDoesNotExist, indexHtmlExists } from '../utils';
+import {
+    t,
+    handleProjectDoesNotExist,
+    indexHtmlExists,
+    determineScpFromTarget,
+    determineUrlFromDestination,
+    determineS4HCFromTarget
+} from '../utils';
 import { getAbapQuestions } from './questions';
 import { EventName } from '../telemetryEvents';
 import { DeployProjectType } from './types';
@@ -27,7 +33,6 @@ import { isInternalFeaturesSettingEnabled } from '@sap-ux/feature-toggle';
 import { isAppStudio } from '@sap-ux/btp-utils';
 import { DEFAULT_PACKAGE_ABAP } from '@sap-ux/abap-deploy-config-inquirer/dist/constants';
 import type { AbapDeployConfig, FioriToolsProxyConfigBackend } from '@sap-ux/ui5-config';
-import type { YeomanEnvironment } from '@sap-ux/fiori-generator-shared';
 import type { AbapDeployConfigOptions } from './types';
 import type {
     AbapDeployConfigAnswersInternal,
@@ -70,12 +75,8 @@ export default class extends DeploymentGenerator {
         await super.initializing();
         await initI18n();
 
-        if ((this.env as unknown as YeomanEnvironment).conflicter) {
-            (this.env as unknown as YeomanEnvironment).conflicter.force = this.options.force ?? true;
-        }
-
         DeploymentGenerator.logger?.debug(t('debug.initTelemetry'));
-        await initTelemetrySettings({
+        await TelemetryHelper.initTelemetrySettings({
             consumerModule: {
                 name: '@sap-ux/abap-deploy-config-sub-generator',
                 version: this.rootGeneratorVersion()
@@ -203,15 +204,30 @@ export default class extends DeploymentGenerator {
             const prompAnswers = await this.prompt(abapDeployConfigPrompts);
             this.answers = reconcileAnswers(prompAnswers, abapAnswers);
         }
-        this._reconcileAnswersWithOptions();
+        await this._reconcileAnswersWithOptions();
     }
 
-    private _processAbapTargetAnswers(): void {
+    private async _processAbapTargetAnswers(): Promise<void> {
         this.answers.destination = this.options.destination || this.answers.destination;
-        this.answers.url = this.options.url || this.answers.url;
+        this.answers.url =
+            this.options.url || this.answers.url || (await determineUrlFromDestination(this.answers.destination));
         this.answers.client = this.options.client || this.answers.client;
-        this.answers.scp = this.options.scp || this.answers.scp;
-        this.answers.isS4HC = this.options.isS4HC || this.answers.isS4HC;
+        this.answers.scp =
+            this.options.scp ||
+            this.answers.scp ||
+            (await determineScpFromTarget({
+                url: this.answers.url,
+                client: this.answers.client,
+                destination: this.answers.destination
+            }));
+        this.answers.isS4HC =
+            this.options.isS4HC ||
+            this.answers.isS4HC ||
+            (await determineS4HCFromTarget({
+                url: this.answers.url,
+                client: this.answers.client,
+                destination: this.answers.destination
+            }));
 
         if (!isAppStudio() && this.answers.scp) {
             // ensure there is no client for SCP on vscode
@@ -234,7 +250,9 @@ export default class extends DeploymentGenerator {
         // Set transport
         if (!this.answers.transport) {
             this.answers.transport =
-                getTransportAnswer(this.options as AbapDeployConfigAnswersInternal) || getTransportAnswer(this.answers);
+                this.options.transport ??
+                (getTransportAnswer(this.options as AbapDeployConfigAnswersInternal) ||
+                    getTransportAnswer(this.answers));
         }
     }
 
@@ -243,8 +261,8 @@ export default class extends DeploymentGenerator {
      *
      * Options may be passed from parent generator, or from the command line.
      */
-    private _reconcileAnswersWithOptions(): void {
-        this._processAbapTargetAnswers();
+    private async _reconcileAnswersWithOptions(): Promise<void> {
+        await this._processAbapTargetAnswers();
         this._processBspAppAnswers();
         this.answers.index = this.options.index ?? this.answers.index;
         this.answers.overwrite = this.options.overwrite ?? this.answers.overwrite;
@@ -252,6 +270,10 @@ export default class extends DeploymentGenerator {
 
     public async writing(): Promise<void> {
         if (!this.launchDeployConfigAsSubGenerator) {
+            await this._writing();
+        } else {
+            // Needed to delay `init` as the yaml configurations won't be ready!
+            await this._initializing();
             await this._writing();
         }
     }
@@ -287,7 +309,7 @@ export default class extends DeploymentGenerator {
     }
 
     public install(): void {
-        if (!this.launchDeployConfigAsSubGenerator && this.answers.overwrite !== false) {
+        if (this.answers.overwrite !== false) {
             this._install();
         }
     }
@@ -310,13 +332,6 @@ export default class extends DeploymentGenerator {
     public async end(): Promise<void> {
         if (this.abort || this.answers.overwrite === false) {
             return;
-        }
-
-        // Delayed process of deploy configuration generation if integrated with app generator
-        if (this.launchDeployConfigAsSubGenerator) {
-            await this._initializing();
-            await this._writing();
-            this._install();
         }
 
         if (