+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/axios-extension
 
+## 1.21.2
+
+### Patch Changes
+
+-   61d4060: use accept header for service generator content request from config
+
 ## 1.21.1
 
 ### Patch Changes