+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/cap-config-writer
 
+## 0.10.6
+
+### Patch Changes
+
+-   Updated dependencies [95a816d]
+    -   @sap-ux/fiori-generator-shared@0.12.6
+
+## 0.10.5
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+
 ## 0.10.4
 
 ### Patch Changes